Update imports and reorganize file structure
