+++ src/main/kotlin/org/testing_survey_creator/dto/SurveyDTO.kt
@@ -1,4 +1,4 @@
-package org.testing_survey_creator.model
+package org.testing_survey_creator.dto
 
 // A DTO (Data Transfer Object) is a design pattern used to transfer data between different layers of an application,
 // without exposing the database entities (Survey, Instruction) directly to the API
